fix(syscall): change all syscall returns from u64 to i64, fix 12 critical bugs
Align syscall return type with Linux (long) by changing all 355 syscall
handler functions from `-> u64` to `-> i64`. This makes error returns
explicit with negative errno instead of relying on two's complement tricks.

Also fix 12 critical bugs found in the second-round code review:

- F02-01: SignalFrame uc pointer uses offset_of! for correct alignment
- F03-01: get_zeroed_page converts phys_to_virt before zeroing
- F04-01: RT enqueue guards against double-enqueue with on_rq check
- F04-02: DL dequeue matches by task pointer only, not mutable deadline
- F05-07: do_waitid correctly produces CLD_KILLED for signal-killed children
- F06-02: Semaphore down() removes CAS loop, returns directly on wakeup
- F06-15: Condvar wait() uses prepare_to_wait to prevent lost-wakeup race
- F09-11: SumGuard declares t6 register clobber in inline asm
- F10-03: socket_close uses Arc::from_raw to reclaim leaked reference
- F10-31: sys_socket removes broken fallback returning internal table index
- F11-01: sys_getdents64 uses copy_to_user instead of manual SUM bit
- F11-02: sys_tgkill validates target existence and permission for sig==0

+++ kernel/src/syscall/dispatch.rs
@@ -26,8 +26,8 @@ fn syscall_get_arguments(regs: &PtRegs) -> SyscallArgs {
 
 /// Set system call return value
 #[inline]
-fn syscall_set_return_value(regs: &mut PtRegs, value: u64) {
-    regs.a0 = value;
+fn syscall_set_return_value(regs: &mut PtRegs, value: i64) {
+    regs.a0 = value as u64;
 }
 
 /// System call entry function
@@ -41,7 +41,7 @@ pub extern "C" fn syscall_handler(regs: &mut PtRegs) {
         crate::process::current_pid(), syscall_no, args[0], args[1], args[2]);
 
     // Dispatch based on system call number (sorted by number)
-    let result: u64 = match syscall_no as u32 {
+    let result: i64 = match syscall_no as u32 {
         // ==================== Linux AIO (NR 0-4) ====================
         0 => memory::sys_io_setup(args),       // io_setup
         1 => memory::sys_io_destroy(args),     // io_destroy
@@ -426,7 +426,7 @@ pub extern "C" fn syscall_handler(regs: &mut PtRegs) {
         _ => {
             crate::println!("syscall: unknown syscall {} (args: {:#x}, {:#x}, {:#x})",
                 syscall_no, args[0], args[1], args[2]);
-            (-errno::ENOSYS) as u64
+            -(errno::ENOSYS as i64)
         }
     };
 
@@ -434,5 +434,5 @@ pub extern "C" fn syscall_handler(regs: &mut PtRegs) {
 
     crate::pr_debug!("syscall: pid={}, nr={}, ret={:#x} ({})",
         crate::process::current_pid(), syscall_no, result,
-        if (result as i64) < 0 { "error" } else { "ok" });
+        if result < 0 { "error" } else { "ok" });
 }
